+++ uhabits-core/src/jvmMain/java/org/isoron/uhabits/core/models/Habit.kt
@@ -83,9 +83,9 @@ data class Habit(
         return if (isNumerical) {
             when (targetType) {
                 NumericalHabitType.AT_LEAST ->
-                    value == Entry.NUMERICAL_AUTO ||
-                        (value >= 0 && value != Entry.SKIP && value / 1000.0 >= targetValue)
-                NumericalHabitType.AT_MOST -> value == Entry.NUMERICAL_AUTO
+                    value == Entry.NUMERICAL_AUTO || (value >= 0 && value / 1000.0 >= targetValue)
+                NumericalHabitType.AT_MOST ->
+                    value == Entry.NUMERICAL_AUTO || (value >= 0 && value / 1000.0 <= targetValue)
             }
         } else {
             value != Entry.NO && value != Entry.UNKNOWN